requirements.txt
django==3.1
psycopg2-binary
djangorestframework
django-filter
python-decouple
Pillow
drf-yasg
twilio
celery
django_celery_results